+++ modules/effects/schematics/ng-add/index.spec.ts
@@ -9,7 +9,7 @@ import {
   getTestProjectPath,
   createWorkspace,
   createAppModuleWithEffects,
-} from '../../../schematics-core/testing';
+} from '@ngrx/schematics-core/testing';
 
 describe('Effects ng-add Schematic', () => {
   const schematicRunner = new SchematicTestRunner(
@@ -35,28 +35,34 @@ describe('Effects ng-add Schematic', () => {
     appTree = await createWorkspace(schematicRunner, appTree);
   });
 
-  it('should update package.json', () => {
+  it('should update package.json', async () => {
     const options = { ...defaultOptions };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const packageJson = JSON.parse(tree.readContent('/package.json'));
 
     expect(packageJson.dependencies['@ngrx/effects']).toBeDefined();
   });
 
-  it('should skip package.json update', () => {
+  it('should skip package.json update', async () => {
     const options = { ...defaultOptions, skipPackageJson: true };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const packageJson = JSON.parse(tree.readContent('/package.json'));
 
     expect(packageJson.dependencies['@ngrx/effects']).toBeUndefined();
   });
 
-  it('should create an effect', () => {
+  it('should create an effect', async () => {
     const options = { ...defaultOptions };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const files = tree.files;
     expect(
       files.indexOf(`${projectPath}/src/app/foo/foo.effects.spec.ts`)
@@ -66,10 +72,12 @@ describe('Effects ng-add Schematic', () => {
     ).toBeGreaterThanOrEqual(0);
   });
 
-  it('should not create an effect if the minimal flag is provided', () => {
+  it('should not create an effect if the minimal flag is provided', async () => {
     const options = { ...defaultOptions, minimal: true };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const files = tree.files;
     const content = tree.readContent(`${projectPath}/src/app/app.module.ts`);
 
@@ -80,54 +88,62 @@ describe('Effects ng-add Schematic', () => {
     expect(files.indexOf(`${projectPath}/src/app/foo/foo.effects.ts`)).toBe(-1);
   });
 
-  it('should not import an effect into a specified module in the minimal flag is provided', () => {
+  it('should not import an effect into a specified module in the minimal flag is provided', async () => {
     const options = {
       ...defaultOptions,
       minimal: true,
       module: 'app.module.ts',
     };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(`${projectPath}/src/app/app.module.ts`);
     expect(content).not.toMatch(
       /import { FooEffects } from '.\/foo\/foo.effects'/
     );
   });
 
-  it('should be provided by default', () => {
+  it('should be provided by default', async () => {
     const options = { ...defaultOptions };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(`${projectPath}/src/app/app.module.ts`);
     expect(content).toMatch(/import { FooEffects } from '.\/foo\/foo.effects'/);
   });
 
-  it('should import into a specified module', () => {
+  it('should import into a specified module', async () => {
     const options = { ...defaultOptions, module: 'app.module.ts' };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(`${projectPath}/src/app/app.module.ts`);
     expect(content).toMatch(/import { FooEffects } from '.\/foo\/foo.effects'/);
   });
 
-  it('should fail if specified module does not exist', () => {
+  it('should fail if specified module does not exist', async () => {
     const options = {
       ...defaultOptions,
       module: `${projectPath}/src/app/app.moduleXXX.ts`,
     };
     let thrownError: Error | null = null;
     try {
-      schematicRunner.runSchematic('effects', options, appTree);
+      await schematicRunner.runSchematicAsync('effects', options, appTree);
     } catch (err) {
       thrownError = err;
     }
     expect(thrownError).toBeDefined();
   });
 
-  it('should respect the skipTests flag', () => {
+  it('should respect the skipTests flag', async () => {
     const options = { ...defaultOptions, skipTests: true };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const files = tree.files;
     expect(
       files.indexOf(`${projectPath}/src/app/foo/foo.effects.ts`)
@@ -137,16 +153,18 @@ describe('Effects ng-add Schematic', () => {
     ).toEqual(-1);
   });
 
-  it('should register the root effect in the provided module', () => {
+  it('should register the root effect in the provided module', async () => {
     const options = { ...defaultOptions };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(`${projectPath}/src/app/app.module.ts`);
 
     expect(content).toMatch(/EffectsModule\.forRoot\(\[FooEffects\]\)/);
   });
 
-  it('should add an effect to the empty array of registered effects', () => {
+  it('should add an effect to the empty array of registered effects', async () => {
     const storeModule = `${projectPath}/src/app/store.module.ts`;
     const options = {
       ...defaultOptions,
@@ -158,13 +176,15 @@ describe('Effects ng-add Schematic', () => {
       'EffectsModule.forRoot([])'
     );
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(storeModule);
 
     expect(content).toMatch(/EffectsModule\.forRoot\(\[FooEffects\]\)/);
   });
 
-  it('should add an effect to the existing registered root effects', () => {
+  it('should add an effect to the existing registered root effects', async () => {
     const storeModule = `${projectPath}/src/app/store.module.ts`;
     const options = {
       ...defaultOptions,
@@ -176,7 +196,9 @@ describe('Effects ng-add Schematic', () => {
       'EffectsModule.forRoot([UserEffects])'
     );
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(storeModule);
 
     expect(content).toMatch(
@@ -184,7 +206,7 @@ describe('Effects ng-add Schematic', () => {
     );
   });
 
-  it('should not add an effect to registered effects defined with a variable', () => {
+  it('should not add an effect to registered effects defined with a variable', async () => {
     const storeModule = `${projectPath}/src/app/store.module.ts`;
     const options = { ...defaultOptions, module: 'store.module.ts' };
     appTree = createAppModuleWithEffects(
@@ -193,13 +215,15 @@ describe('Effects ng-add Schematic', () => {
       'EffectsModule.forRoot(effects)'
     );
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const content = tree.readContent(storeModule);
 
     expect(content).not.toMatch(/EffectsModule\.forRoot\(\[FooEffects\]\)/);
   });
 
-  it('should group within an "effects" folder if group is set', () => {
+  it('should group within an "effects" folder if group is set', async () => {
     const options = {
       ...defaultOptions,
       flat: true,
@@ -207,7 +231,9 @@ describe('Effects ng-add Schematic', () => {
       group: true,
     };
 
-    const tree = schematicRunner.runSchematic('ng-add', options, appTree);
+    const tree = await schematicRunner
+      .runSchematicAsync('ng-add', options, appTree)
+      .toPromise();
     const files = tree.files;
     expect(
       files.indexOf(`${projectPath}/src/app/effects/foo.effects.ts`)