+++ modules/schematics/migrations/6_0_0/index.spec.ts
@@ -7,11 +7,11 @@ import * as path from 'path';
 import {
   createPackageJson,
   packagePath,
-} from '../../../schematics-core/testing/create-package';
+} from '@ngrx/schematics-core/testing/create-package';
 import {
   upgradeVersion,
   versionPrefixes,
-} from '../../../schematics-core/testing/update';
+} from '@ngrx/schematics-core/testing/update';
 
 const collectionPath = path.join(__dirname, '../migration.json');
 
@@ -20,16 +20,14 @@ describe('Schematics Migration 6_0_0', () => {
   const pkgName = 'schematics';
 
   versionPrefixes.forEach((prefix) => {
-    it(`should install version ${prefix}6.0.0`, () => {
+    it(`should install version ${prefix}6.0.0`, async () => {
       appTree = new UnitTestTree(Tree.empty());
       const runner = new SchematicTestRunner('schematics', collectionPath);
       const tree = createPackageJson(prefix, pkgName, appTree);
 
-      const newTree = runner.runSchematic(
-        `ngrx-${pkgName}-migration-01`,
-        {},
-        tree
-      );
+      const newTree = await runner
+        .runSchematicAsync(`ngrx-${pkgName}-migration-01`, {}, tree)
+        .toPromise();
       const pkg = JSON.parse(newTree.readContent(packagePath));
       expect(pkg.dependencies[`@ngrx/${pkgName}`]).toBe(
         `${prefix}${upgradeVersion}`